[Refactoring] Rename project to ETL #58
