+++ README.md
@@ -1,4 +1,4 @@
-# Lunches Actualizer
+# Lunches ETL
 
 Actualization - is the process of bringing of data from one place in accordance with the state of another place
 
@@ -15,7 +15,7 @@ This project contains data actualizers and other tools to maintain state of diff
 
 - `$ cd /path/to/project`
 - clone repository
-- `$ cd lunches-actualizer`
+- `$ cd lunches-etl`
 - `$ curl -sS https://getcomposer.org/installer | php`
 - `$ php composer.phar install`
 - `$ cp parameters.php.dist parameters.php`
